Formatting code

+++ src/main/java/pl/coderslab/users/model/UserDelete.java
@@ -1,11 +1,11 @@
 package pl.coderslab.users.model;
-
 import pl.coderslab.users.controller.User;
 import pl.coderslab.users.controller.UserDAO;
 
 import javax.servlet.*;
 import javax.servlet.http.*;
 import javax.servlet.annotation.*;
+import javax.swing.*;
 import java.io.IOException;
 
 @WebServlet("/user/delete")
@@ -22,9 +22,13 @@ public class UserDelete extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         int id = Integer.parseInt(request.getParameter("id"));
-        UserDAO userDAO = new UserDAO();
-        userDAO.delete(id);
-        response.sendRedirect(request.getContextPath() + "/user/list");
-
+        int option = JOptionPane.showConfirmDialog(null, "Are you sure you want to delete?", "Delete", JOptionPane.YES_NO_OPTION);
+        if (option == 0) {
+            UserDAO userDAO = new UserDAO();
+            userDAO.delete(id);
+            response.sendRedirect(request.getContextPath() + "/user/list");
+        } else {
+            response.sendRedirect(request.getContextPath() + "/user/list");
+        }
     }
 }
